pyproject.toml
[tool.poetry]
name = "users-crud-api"
version = "0.1.0"
description = "API RESTful de CRUD de Usuários com FastAPI"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.0"
uvicorn = {version = "^0.27.0", extras = ["standard"]}
sqlalchemy = "^2.0.25"
aiosqlite = "^0.20.0"
pydantic = "^2.8.2"
pydantic-settings = "^2.3.0"
email-validator = "^2.3.0"
greenlet = "^3.5.0"
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
bcrypt = "^4.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.0"
httpx = "^0.27.0"
ruff = "^0.4.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
